// src/app/api/models/CursusDtoIEnumerableResponseDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CursusDto } from './CursusDto';
export type CursusDtoIEnumerableResponseDTO = {
    message?: string | null;
    status?: number | null;
    data?: Array<CursusDto> | null;
    count?: number | null;
};