// src/app/api/models/CursusDtoResponseDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CursusDto } from './CursusDto';
export type CursusDtoResponseDTO = {
    message?: string | null;
    status?: number | null;
    data?: CursusDto;
    count?: number | null;
};